+++ src/features/getProfile/model/index.ts
@@ -1,0 +1,2 @@
+export * from './useProfileQuery';
+export * from './useProfileStore';
