+++ src/features/getProfile/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as Profile } from './Profile';
