+++ src/widgets/sidebar/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as SideBar } from './SideBar';
